fix(api): count(*) retorna number + 4 fixes preexistentes
count(*)::int aplicado en 18 endpoints de listado para que meta.total
retorne number en lugar de string (postgres mapea bigint→string por
defecto). Resuelve fails de integration tests en time-windows, skills,
misc/presets, vehicle-skills.

Otros fixes incluidos:
- /api/mobile/driver/location GET: requireRoutePermission(DRIVER:READ)
  era contradictorio (CONDUCTOR no tiene ese permiso). Ahora solo exige
  auth y la lógica posterior (que ya filtraba por self) hace el role
  check.
- 3 tests company-profiles: actualizados de 401→400 (ADMIN_SISTEMA sin
  x-company-id devuelve 400 COMPANY_REQUIRED, no 401).
- plan-confirmation test: actualizado para reflejar que el handler
  permite confirmar desde DRAFT o CONFIGURED (no solo DRAFT).

Backlog (no fixeados, issues independientes):
- companies/workflow: tenant isolation regresión (403 vs 200)
- companies/management: ADMIN_FLOTA cross-company (401 vs 403)
- orders/import: 9 tests del módulo CSV import roto
- monitoring: beforeEach timeout
- optimization: DELETE COMPLETED job 400 vs ?

+++ src/app/api/companies/route.ts
@@ -61,7 +61,7 @@ export async function GET(request: NextRequest) {
         .orderBy(desc(companies.createdAt))
         .limit(query.limit)
         .offset(query.offset),
-      db.select({ count: sql<number>`count(*)` }).from(companies).where(whereClause),
+      db.select({ count: sql<number>`count(*)::int` }).from(companies).where(whereClause),
     ]);
 
     return NextResponse.json({
